# Optimize word mover distance (WMD) computation