@mpenkov 2021-06-29
Merged. Congrats on your first PR to gensim @flowlight0 ! 🥇