@flowlight0 2021-06-29
Thanks for the gensim team's effort on reviewing and merging this PR!